TMS3203x core updates:  * fixed interrupt handling  * added support for edge-triggered interrupts on '32  * expanded interrupt support for the '32  * updated drivers using TMS3203x core to deassert interrupts  * added externally accessible functions for converting '3x floating point format  * updated gaelco3d driver to use new functions
Zeus2 (+related) updates:
 * fixed save states for DCS games
 * cleaned up Zeus2 waveram handling
 * added Zeus2 save state support
 * added preliminary model and quad rendering support for Zeus2
 * added support to timekpr for the ZPRAM used on Zeus2
 * hooked up ZPRAM in Zeus2 games
 * hooked up controls in Zeus2 games
 * updated poly.c to ensure it is idle before saving state

+++ src/mame/video/midzeus.c
@@ -1392,7 +1392,7 @@ static void log_waveram(UINT32 length_and_base)
 
 	UINT32 numoctets = (length_and_base >> 24) + 1;
 	const UINT32 *ptr = waveram_ptr_from_block_addr(0, length_and_base);
-	UINT64 checksum = length_and_base;
+	UINT32 checksum = length_and_base;
 	int foundit = FALSE;
 	int i;
